requirements.txt
# HunyuanOCR
# Note: transformers need to be installed from specific commit for HunyuanOCR support
# pip install git+https://github.com/huggingface/transformers@82a06db03535c49aa987719ed0746a76093b1ec4
transformers>=4.37.0
pillow>=10.0.0
accelerate>=0.20.0
opencv-python>=4.8.0
sentencepiece

# PyTorch - supports CUDA, MPS (Apple Silicon), and CPU
# For CUDA (NVIDIA GPU):
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# 
# For CPU-only or MPS (Mac):
# pip install torch torchvision torchaudio
#
# When installing from requirements.txt:
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
peft>=0.18.1